Handle TempLanguage save; Set User relations to uneditable
Create a post_save handler for TempLanguage so that creating a new
TempLanguage will also create a Language and link them together.

Activities (mainly creating Charters and Events) can continue as usual
using this newly created Language. Relationships and history between
TempLanguage, Language, Charter, and Event will remain unchanged when
the time comes to give Language a permanent, or real, IETF tag.

------------------------------------------------------------------------

Set created_by and modified_by of TempLanguage to be uneditable. They
will be manually set when saving through the ModelForm or through the
Admin interface. It makes sense to put these out of human reach just
like created_at and modified_at

------------------------------------------------------------------------

Modify ResourceDocument.language to SET_NULL on delete as part of
figuring out a migration error that I experience. Feel free to change it
back if it's supposed to be on CASCADE mode.