Add UpdateView, mapping classmethods to TempLanguage
Create UpdateView for TempLanguage using the same template as
CreateView, but with a couple test condition to display the appropriate
title and disable IETF tag editing.

------------------------------------------------------------------------

Add lang_assigned_map() and lang_assigned_changed_map() as classmethod
of TempLanguage to see to what language is the temporary language
assigned. This info could be useful for offline devices that need to
update their database.