Switch TempLanguage CreateView to WizardView
To display the questions in a more user-friendly manner, it's good to
not show all the questions in the questionnaire at the same time.
Breaking the questions into several parts (determined by the
'depends_on' attr.) helps readability. This change will work in tandem
with a well-crafted questionnaire to offer a good user experience.