Create mapping from answer to model fields
The automatically-generated Language object (by TempLanguage) needs to
have important identifying information: name and country (and region)
for future referencing. By adding 'field_mapping' to Questionnaire, we
would allow the creator of the questionnaire to specify for which field
in TempLanguage model is an answer.

Update the information displayed in the Temporary Language Admin and
email notification to reflect the changes mentioned above.