Add ability to create a temporary language with questionnaire
* Re-structure homepage and static/src

Create a new section: "Searching for a Language" that acts as a
shortcut to search the language list.

Majority of the use-cases are user searching for a language. It makes
sense to provide a way to do that straight from the home screen.

Re-position and re-style main sections in the home page to
accomodate a new section: "Searching for a language"

Re-position elements that make up the GL map so that they rely on
the actual height of the map itelf rather than an arbitrary % value
of padding-bottom property.

The current GL map styling is out of order. This is noticeable because
the PDF download button is misaligned on the page. Further CSS
inspection reveals problems with how the <svg> is generated and
positioned.

Refactor current static/src/js and static/src/less folder to allow
modularization on the developer side.

Currently, main.js and site.less contain almost all the content for the
whole site. That makes it difficult to find codes specific to a page.
The refactoring helps to store specific codes in a specific file.

main.js and site.less act as the primary files. All other files should
be imported to them. See static/src for the convention.

* Create TempLanguage model

* Register TempLanguage in admin

* Handle TempLanguage save; Set User relations to uneditable

Create a post_save handler for TempLanguage so that creating a new
TempLanguage will also create a Language and link them together.

Activities (mainly creating Charters and Events) can continue as usual
using this newly created Language. Relationships and history between
TempLanguage, Language, Charter, and Event will remain unchanged when
the time comes to give Language a permanent, or real, IETF tag.

Set created_by and modified_by of TempLanguage to be uneditable. They
will be manually set when saving through the ModelForm or through the
Admin interface. It makes sense to put these out of human reach just
like created_at and modified_at

Modify ResourceDocument.language to SET_NULL on delete as part of
figuring out a migration error that I experience. Feel free to change it
back if it's supposed to be on CASCADE mode.

* Add basic UI and View for getting temporary code

* Switch TempLanguage form to CreateView and ModelForm

Utilize ModelForm to create TempLanguage instead of the previous basic
UI. The basic UI would have become a form eventually.

* Add TempLanguage List-, DetailView; Improve creation workflow

Create TempLanguage ListView and DetailView to view existing temporary
languages. There is no current way to confirm TempLanguage creation on
the front-end.

Improve the workflow of creating TempLanguage by applying the following
changes:
 - Register unique constrain to the TempLanguage.ietf_tag so that no
duplicate temporary code can be created.
 - Change redirect to TempLanguage detail page after successfully
creating an instance.
 - Notify interested parties via email when a temporary language is
created

* Add UpdateView, mapping classmethods to TempLanguage

Create UpdateView for TempLanguage using the same template as
CreateView, but with a couple test condition to display the appropriate
title and disable IETF tag editing.

Add lang_assigned_map() and lang_assigned_changed_map() as classmethod
of TempLanguage to see to what language is the temporary language
assigned. This info could be useful for offline devices that need to
update their database.

* Create admin view for TempLanguage review process

Create a simple TemplateView that displays pending, approved, or
rejected temporary languages. This is intended to be used by Ralph
Sievert and his team, or whoever is tasked with reviewing temporary
language submissions. Still needs to be wired up.

* Create Questionnaire and update TempLanguage

Create Questionnaire model to hold questionnaire questions created by
Ralph and/or his team. Making a table for the questionnaire itself
allows us to "version control" the questions, allowing submission for
older, or oudated, questionnaire.

TempLanguage is going to act as the request ticket, therefore it is
updated with necessary fields to hold extra information about the
request. Some unnecessary fields are deleted because the info can be
inferred from the answer to the questionnaire attached to it.

* Set Language ordering by name

* Add migration file for language ordering

* Return a string not an int

* Fixed a couple things that broke with model change

* Modified the details page to remove/add fields

* Fixed a spelling issue

* Display questions and answers

* Switch TempLanguage CreateView to WizardView

To display the questions in a more user-friendly manner, it's good to
not show all the questions in the questionnaire at the same time.
Breaking the questions into several parts (determined by the
'depends_on' attr.) helps readability. This change will work in tandem
with a well-crafted questionnaire to offer a good user experience.

* Fix flake8 errors

* Some finishing touches on the detail page

* Initial tests

* Add unittest for TempLanguage

* Change from 'answer' to 'text' and fix a flake8 issue

* Always use strings for matching

* Tests finished

* Added entry points for the temp language form

* Build basic TempLanguage approval process

* Fix multiple creation of Language on TempLanguage handler

When a TempLang is approved (by changing the Lang code and saving), the
TempLang post_save handler couldn't find the language with the TempLang
code and therefore creates another new Language. The solution is to
check, in the handler, if it's run because of creation or modification.
Only creates a Language object if TempLang is a newly created object.

* Redirect to newly created language instead of the temp language

* Assign required_css_class for the forms in the questionnaire

* Change TempLang Q&A to use ordinal numbering

This is for the one in the TempLanguage detail page.

* Fixing flake8 issues

* Create mapping from answer to model fields

The automatically-generated Language object (by TempLanguage) needs to
have important identifying information: name and country (and region)
for future referencing. By adding 'field_mapping' to Questionnaire, we
would allow the creator of the questionnaire to specify for which field
in TempLanguage model is an answer.

Update the information displayed in the Temporary Language Admin and
email notification to reflect the changes mentioned above.

* Move TempLanguageWizardView.questionnaire into __init__

A migration error is possibly caused by defining questionnaire directly
under TempLanguageWizardView.

* Add some unit-tests

* Condense TempLanguage and Questionnaire migration files

* Fix flake8 errors

* Add unit tests for TempLanguage views

* Adjust changes based on review feedback

* Revert changes to ResourceDocument; Notify ietf@unfoldingword.org