Add word count as weight to Documents
The progress rate for each GL is skewed because some documents are
bigger/smaller than the others. 100% completion of a 200-word document
is not the same as 100% completion of a 30,000-word document.

Word-count is added as a way to weigh these documents. Therefore, 100%
completion rate for a 30,000-word document is significantly heftier
than the one for a 200-word document. The reverse is also true.

Also refactored PhaseView and its functions while at it.