Add new_start field to Charter
Users need a way to identify that a charter is created for a brand new
translation project for the language, as opposed to continuing/joining
an effort initiated by other party.

Styles for the pages touched by this change are moved to its own file as
part of the continuous effort to structure the project.