Make minor adjustments based on review feedback
Event-count table is now also available in Events dashboard.

The selection for fiscal year is expended beyond "this", "next", and
"last". Now it displays the years from current minus one to current
plus four.

As a side effect to the above change, the url pattern for
AjaxEventCountView is adjusted to allow other numbers besides 0, 1, and
-1.

Fix the back-end pagination to allow datatables to show "All" contents,
which is noted by a negative integer (-1).