Customize Tag from django-taggit
We need a Tag model/table to be some sort of "tag register" that lists
models (WARegion, Country, Language, Charter, Event are intendended)
alongside their tag slug. In other words, it need to keys to one of the
models in addition to the comment.

In admin, we'll use this custom tag (CommentTag) instead of
django-taggit's Tag.