Commenting system phase 1 (#608)
* Initial integration of django_comments and django-taggit

* Customize Tag from django-taggit

We need a Tag model/table to be some sort of "tag register" that lists
models (WARegion, Country, Language, Charter, Event are intendended)
alongside their tag slug. In other words, it need to keys to one of the
models in addition to the comment.

In admin, we'll use this custom tag (CommentTag) instead of
django-taggit's Tag.

* Add tag_slug property to WARegion, Country, Language, Charter, and Event

* Create management command to update CommentTag

* Add receivers and handler function to update comment tag

* Add hashtag info to WARegion, Country, Language, Charter, and Event detail page

* Parse hashtags in comments

* Add comment sections to Country, Language, and Event

* Add commenting to charter tab under language

* Add filter for all, originals, and mentions

* Add commenting management commands

* Display object name instead of the tag slug

* Prep work for review

Adjust tag slugging rule to use language's ISO-639-3 instead of its
code (IETF) if the code has less than 3 letters.

Fix duplicate comments and comment counts when writing a comment that
mentions itself.

Add missing "mentions" property to Charter and Event.

* Let comments and mentions be called as an object property.

Abstract out tags and comments properties in CommentableModel.

* Put events' comments under Translation Work tab in Language detail

* Make post_comment a class-based view

* Fix broken unit-tests

IntegrityError in create_comment_tag() causes database to close
connection early. Hence breaking the rest of the unit-tests.

* Add ability to delete one's own comment

* Write unit-tests